requirements.txt
flask
pymongo
python-docx
requests
gunicorn
huggingface_hub